Titanic-kaggle-compitetion-main/Titanic-kaggle-compitetion-main/Titanic Compitetion/requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
optuna
shap
mlflow
evidently
fastapi
uvicorn
streamlit
pyyaml
hydra-core
pytest
matplotlib
seaborn
pydantic
python-dotenv
joblib
